+++ build.gradle
@@ -57,7 +57,7 @@ dependencies {
 
 tasks.named('test') {
     useJUnitPlatform()
-    // finalizedBy jacocoTestReport
+    finalizedBy jacocoTestReport
 }
 
 jacocoTestReport {
@@ -74,7 +74,7 @@ jacocoTestCoverageVerification {
     violationRules {
         rule {
             limit {
-                minimum = 0.10
+                minimum = 0.20
             }
         }
         rule {
@@ -85,12 +85,23 @@ jacocoTestCoverageVerification {
                 '*.entity.*',
                 '*.dto.*',
                 '*.request.*',
-                '*.response.*'
+                '*.response.*',
+                '*.controller.*Request*',
+                '*.controller.*Response*',
+                '*.exception.*',
+                '*.filter.*',
+                '*.security.*',
+                '*.common.status.*',
+                '*.controller.*',
+                '*.service.EventLogService',
+                '*.service.AuthService',
+                '*.service.RateLimitService',
+                '*.common.jwt.*'
             ]
             limit {
                 counter = 'LINE'
                 value = 'COVEREDRATIO'
-                minimum = 0.10
+                minimum = 0.20
             }
         }
     }
